+++ src/components/scroll-effects-control/effect-properties.js
@@ -60,9 +60,37 @@ const EFFECT_PROPERTIES = {
 		max: 360,
 		unitLabel: 'deg',
 	},
+	rotateX: {
+		label: 'Angle',
+		attr: 'rotateX',
+		min: -360,
+		max: 360,
+		unitLabel: 'deg',
+	},
+	rotateY: {
+		label: 'Angle',
+		attr: 'rotateY',
+		min: -360,
+		max: 360,
+		unitLabel: 'deg',
+	},
 	scale: {
 		label: 'Scale',
 		attr: 'scale',
+		min: 0,
+		max: 1000,
+		unitLabel: '%',
+	},
+	scaleX: {
+		label: 'Scale',
+		attr: 'scaleX',
+		min: 0,
+		max: 1000,
+		unitLabel: '%',
+	},
+	scaleY: {
+		label: 'Scale',
+		attr: 'scaleY',
 		min: 0,
 		max: 1000,
 		unitLabel: '%',
